requirements.txt
discord.py
selenium
undetected-chromedriver
